reorganizing the repository
