+++ .gitignore
@@ -1,3 +1,6 @@
+data/
+spikeinterface/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
