--- spikeinterface/deepinterpolation_recording/trained_models/test_training_bp_filter_t20s_v0.5s/generator.json
@@ -1,1 +1,0 @@
-{"type": "generator", "name": "EphysGenerator", "pre_post_frame": 30, "pre_post_omission": 1, "steps_per_epoch": -1, "train_path": "/home/buccino/codes/deepinterpolation-tutorial/deepinterpolation/examples/../../../../data/Neuropixels2.0_Recording/open-ephys-np2/595262_2022-02-22_16-47-26/Record Node 102/experiment1/recording1/continuous/Neuropix-PXI-100_ProbeA-AP/continuous.dat", "batch_size": 100, "start_frame": 10000000, "end_frame": 10001000, "randomize": 0}
