--- spikeinterface/deepinterpolation_recording/trained_models/test_training_bp_filter_t20s_v0.5s/inferrence.json
@@ -1,1 +1,0 @@
-{"type": "inferrence", "name": "core_inferrence", "model_path": "/home/buccino/codes/ephys-deepinterpolation/spikeinterface/test_training_bp_filter_t20s_v0.5s/first_test_unet_single_ephys_1024_mean_absolute_error_model.h5", "output_file": "/home/buccino/codes/ephys-deepinterpolation/spikeinterface/test_training_bp_filter_t20s_v0.5s/inference_test_training_bp_filter_t20s_v0.5s.h5"}
