--- spikeinterface/deepinterpolation_recording/trained_models/test_training_bp_filter_t60_v0.5s_assertion/network_params.json
@@ -1,1 +1,0 @@
-{"type": "network", "name": "unet_single_ephys_1024"}
