--- spikeinterface/deepinterpolation_recording/trained_models/test_training_hp_filter_t20s_v0.5s/training_params.json
@@ -1,1 +1,0 @@
-{"loss": "mean_absolute_error", "model_string": "unet_single_ephys_1024_mean_absolute_error", "output_dir": "test_training_hp_filter_t20s_v0.5s", "run_uid": "first_test", "nb_gpus": 1, "type": "trainer", "steps_per_epoch": 10, "period_save": 5, "apply_learning_decay": 0, "nb_times_through_data": 1, "learning_rate": 0.0001, "pre_post_frame": 1, "nb_workers": 2, "caching_validation": false}
