--- spikeinterface/deepinterpolation_recording/trained_models/test_training_no_hp_filter_t20s_v0.5s/network_params.json
@@ -1,1 +1,0 @@
-{"type": "network", "name": "unet_single_ephys_1024"}
