--- toy_example_files/visualization/Readme.txt
@@ -1,9 +1,0 @@
-**Visualize Output of Inference**
-In order to visualize the figures, the python code file "Visualize Content of Inference" should be inserted in the same file containing the model.
-A figure is thus generated showing the denoised (inferred) results.
-
-**Visualize Content of Data Sample**
-In order to visualize the figures, the python code file "Visualize Content of Data Sample (.dat2) File" should be inserted in the same file containing the raw dataset.
-A figure is thus generated showing the original raw data.
-
-Note: for scaling purposes, and for showing the output in microVolts, the data can be multiplied by .195 (the code line for this is present in both files and is currently supressed)..
